pyproject.toml
[project]
name = "selenium-ide-ai-side-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "openai>=2.24.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "types-click>=7.1.8",
]
